+++ src/db/MetadataEntity.ts
@@ -1,4 +1,4 @@
-import {Column, CreateDateColumn, DeleteDateColumn, PrimaryGeneratedColumn, UpdateDateColumn} from "typeorm";
+import {CreateDateColumn, DeleteDateColumn, PrimaryGeneratedColumn, UpdateDateColumn} from "typeorm";
 
 export abstract class MetadataEntity {