+++ src/service/auth/auth-role.enum.ts
@@ -1,5 +1,0 @@
-export enum AuthRole {
-    ADMIN = '2',
-    USER = '1',
-    GUEST = '0',
-}