+++ src/web/Endpoint.interface.ts
@@ -1,10 +1,8 @@
-import {HttpVerb} from "./http-verb.enum";
+import {HttpVerb} from "./HttpVerb.enum";
 import express from "express";
-import {AuthRole} from "../service/auth/auth-role.enum";
 
 export interface Endpoint {
     verb: HttpVerb,
     path: string,
     handler: express.RequestHandler,
-    minimalRole?: AuthRole
 }