+++ test/tests.utils.ts
@@ -1,6 +1,5 @@
 import {DataSourceOptions} from "typeorm";
 import path from "path";
-import jwt from "jsonwebtoken";
 
 export abstract class TestsUtils {
 
@@ -10,21 +9,4 @@ export abstract class TestsUtils {
         synchronize: true,
         entities: [path.resolve(__dirname, '..') + '/**/*.entity{.ts,.js}'],
     };
-
-    public static testUser = {
-        username: 'username',
-        password: 'password',
-        roles: [
-            {name: 'admin'},
-        ]
-    }
-
-    public static getValidAccessToken(): string {
-        const payload = {
-            sub: '1',
-            username: 'username',
-            role: 'admin',
-        };
-        return jwt.sign(payload, 'ACCESS SECRET', {expiresIn: '1h'});
-    }
 }